pupa conversion of legislators.py
